test: ut assertions move to nut
