+++ test/commands/dev/generate/plugin.nut.ts
@@ -5,9 +5,10 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import path from 'node:path';
+import { existsSync, readFileSync } from 'node:fs';
 import { TestSession, execInteractiveCmd, Interaction } from '@salesforce/cli-plugins-testkit';
 import { expect } from 'chai';
-import { fileExists, readJson } from '../../../../src/util.js';
+import { readJson } from '../../../../src/util.js';
 import { NYC, PackageJson } from '../../../../src/types.js';
 
 describe('dev generate plugin NUTs', () => {
@@ -31,9 +32,10 @@ describe('dev generate plugin NUTs', () => {
       },
       { cwd: session.dir, ensureExitCode: 0 }
     );
+    const pluginDir = path.join(session.dir, 'plugin-awesome');
 
-    const packageJsonPath = path.join(session.dir, 'plugin-awesome', 'package.json');
-    expect(await fileExists(packageJsonPath)).to.be.true;
+    const packageJsonPath = path.join(pluginDir, 'package.json');
+    expect(existsSync(packageJsonPath)).to.be.true;
 
     const packageJson = readJson<PackageJson>(packageJsonPath);
 
@@ -43,6 +45,25 @@ describe('dev generate plugin NUTs', () => {
     expect(packageJson.bugs).to.equal('https://github.com/forcedotcom/cli/issues');
     expect(packageJson.repository).to.equal('salesforcecli/plugin-awesome');
     expect(packageJson.homepage).to.equal('https://github.com/salesforcecli/plugin-awesome');
+
+    const testDependencyScripts = packageJson.wireit.test.dependencies;
+    expect(testDependencyScripts).to.include('test:json-schema');
+    expect(testDependencyScripts).to.include('test:deprecation-policy');
+    expect(testDependencyScripts).to.include('test:command-reference');
+
+    // assert files created
+    [
+      'CODE_OF_CONDUCT.md',
+      'LICENSE.txt',
+      'command-snapshot.json',
+      path.join('schemas', 'hello-world.json'),
+      path.join('.git2gus', 'config.json'),
+    ].map((file) => {
+      expect(existsSync(path.join(pluginDir, file)), `${file} does not exist!`).to.be.true;
+    });
+
+    expect(readFileSync(path.join(pluginDir, 'src', 'commands', 'hello', 'world.ts'), 'utf8')).to.include('Copyright');
+    expect(readFileSync(path.join(pluginDir, '.eslintrc.cjs'), 'utf8')).to.include('eslint-config-salesforce-license');
   });
 
   it('should generate a 3PP plugin', async () => {
@@ -57,9 +78,9 @@ describe('dev generate plugin NUTs', () => {
       },
       { cwd: session.dir, ensureExitCode: 0 }
     );
-
-    const packageJsonPath = path.join(session.dir, 'my-plugin', 'package.json');
-    expect(await fileExists(packageJsonPath)).to.be.true;
+    const pluginDir = path.join(session.dir, 'my-plugin');
+    const packageJsonPath = path.join(pluginDir, 'package.json');
+    expect(existsSync(packageJsonPath)).to.be.true;
 
     const packageJson = readJson<PackageJson>(packageJsonPath);
 
@@ -67,7 +88,29 @@ describe('dev generate plugin NUTs', () => {
     expect(packageJson.author).to.equal('me');
     expect(packageJson.description).to.equal('a description');
 
-    const nycConfig = readJson<NYC>(path.join(session.dir, 'my-plugin', '.nycrc'));
+    const scripts = Object.keys(packageJson.scripts);
+    expect(scripts).to.not.include('test:json-schema');
+    expect(scripts).to.not.include('test:deprecation-policy');
+    expect(scripts).to.not.include('test:command-reference');
+
+    const keys = Object.keys(packageJson);
+    expect(keys).to.not.include('homepage');
+    expect(keys).to.not.include('repository');
+    expect(keys).to.not.include('bugs');
+
+    // assert files created
+    expect(existsSync(path.join(pluginDir, 'src', 'commands', 'hello', 'world.ts'))).to.be.true;
+
+    // assert files not created
+    ['CODE_OF_CONDUCT.md', 'LICENSE.txt', 'command-snapshot.json', 'schemas', '.git2gus'].map((file) => {
+      expect(existsSync(path.join(pluginDir, file))).to.be.false;
+    });
+
+    expect(readFileSync(path.join(pluginDir, 'src', 'commands', 'hello', 'world.ts'), 'utf8')).to.not.include(
+      'Copyright'
+    );
+
+    const nycConfig = readJson<NYC>(path.join(pluginDir, '.nycrc'));
     expect(nycConfig).to.deep.equal({
       'check-coverage': true,
       lines: 75,
@@ -92,14 +135,13 @@ describe('dev generate plugin NUTs', () => {
     );
 
     const packageJsonPath = path.join(session.dir, 'my-plugin', 'package.json');
-    expect(await fileExists(packageJsonPath)).to.be.true;
+    expect(existsSync(packageJsonPath)).to.be.true;
 
     const packageJson = readJson<PackageJson>(packageJsonPath);
 
     expect(packageJson.name).to.equal('my-plugin');
     expect(packageJson.author).to.equal('me');
     expect(packageJson.description).to.equal('a description');
-
     const nycConfig = readJson<NYC>(path.join(session.dir, 'my-plugin', '.nycrc'));
     expect(nycConfig).to.deep.equal({
       'check-coverage': true,
